refactor autoupdate-module
