+++ ot-node.js
@@ -3,7 +3,7 @@ const DeepExtend = require('deep-extend');
 const rc = require('rc');
 const fs = require('fs');
 const queue = require('fastq');
-const AutoUpdater = require('./modules/auto-update/auto-updater-module-interface');
+const AutoUpdater = require('./src/modules/auto-updater/auto-updater-module-interface');
 const DependencyInjection = require('./modules/service/dependency-injection');
 const Logger = require('./modules/logger/logger');
 const constants = require('./modules/constants');
