+++ src/modules/auto-updater/auto-updater-module-interface.js
