+++ src/modules/auto-updater/implementation/ot-auto-updater.js
